Gallery
Using at least 20 images. Develop an image gallery. The web page must:

Be developed using Flexbox layout showed in class
Add an icon on the corner of each image to toggle between ON/OFF
ON means that image is a favorite image and it will be bigger
OFF means it is a non-favorite image.
On the top of the gallery, have a button "add image"
Not don't necessarily have button element, it's up to you what to use, it just need to be clickable.
This button should not scroll with the gallery
When clicked, it should allow the user to add an image into the library by providing a URL.